+++ apps/fluent-tester/src/E2E/Avatar/specs/Avatar.spec.win.ts
@@ -2,9 +2,9 @@ import NavigateAppPage from '../../common/NavigateAppPage.win';
 import AvatarPageObject from '../pages/AvatarPageObject';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 import {
-  JSAVATAR_ACCESSIBILITY_LABEL,
-  JSAVATAR_ACCESSIBILITY_LABEL_BY_NAME,
-  JSAVATAR_ACCESSIBILITY_HINT,
+  AVATAR_ACCESSIBILITY_LABEL,
+  AVATAR_ACCESSIBILITY_LABEL_BY_NAME,
+  AVATAR_ACCESSIBILITY_HINT,
   ACCESSIBILITY_LABEL_ATTR,
   ACCESSIBILITY_HINT_ATTRIBUTE,
   ACCESSIBILITY_ROLE_ATTRIBUTE,
@@ -42,13 +42,13 @@ describe('Avatar Accessibility Testing', () => {
     expect(AvatarPageObject.didAssertPopup()).toBeFalsy(AvatarPageObject.ERRORMESSAGE_ASSERT);
   });
   it('Validate accessibilityLabel', () => {
-    expect(AvatarPageObject.getPrimaryComponentAttribute(ACCESSIBILITY_LABEL_ATTR)).toEqual(JSAVATAR_ACCESSIBILITY_LABEL);
+    expect(AvatarPageObject.getPrimaryComponentAttribute(ACCESSIBILITY_LABEL_ATTR)).toEqual(AVATAR_ACCESSIBILITY_LABEL);
   });
   it('Validate accessibilityLabel from `name` prop', () => {
-    expect(AvatarPageObject.getSecondaryComponentAttribute(ACCESSIBILITY_LABEL_ATTR)).toEqual(JSAVATAR_ACCESSIBILITY_LABEL_BY_NAME);
+    expect(AvatarPageObject.getSecondaryComponentAttribute(ACCESSIBILITY_LABEL_ATTR)).toEqual(AVATAR_ACCESSIBILITY_LABEL_BY_NAME);
   });
   it('Validate accessibilityHint', () => {
-    expect(AvatarPageObject.getPrimaryComponentAttribute(ACCESSIBILITY_HINT_ATTRIBUTE)).toEqual(JSAVATAR_ACCESSIBILITY_HINT);
+    expect(AvatarPageObject.getPrimaryComponentAttribute(ACCESSIBILITY_HINT_ATTRIBUTE)).toEqual(AVATAR_ACCESSIBILITY_HINT);
   });
   it('Validate accessibilityRole', () => {
     expect(AvatarPageObject.getPrimaryComponentAttribute(ACCESSIBILITY_ROLE_ATTRIBUTE)).toEqual(ACCESSIBILITY_ROLE_LINK);