+++ apps/fluent-tester/src/FluentTester/TestComponents/Avatar/BasicAvatar.tsx
@@ -4,10 +4,10 @@ import {
   AvatarSizes,
   AvatarColor,
   AvatarColors,
-  JSAvatar,
+  Avatar,
   AvatarActive,
   AvatarActiveAppearance,
-} from '@fluentui-react-native/experimental-avatar';
+} from '@fluentui-react-native/avatar';
 import { PresenceBadgeStatuses, PresenceBadgeStatus } from '@fluentui-react-native/badge';
 import { Switch, View, Text, Picker, ColorValue, Platform } from 'react-native';
 import { satyaPhotoUrl, undefinedText } from './../PersonaCoin/styles';
@@ -68,7 +68,6 @@ export const StandardUsage: FunctionComponent = () => {
     src: TestSvg,
     viewBox: '0 0 500 500',
   };
-  const iconProps = { svgSource: svgProps, width: 20, height: 20, color: 'red' };
 
   const svgIconsEnabled = ['ios', 'macos', 'win32', 'android'].includes(Platform.OS as string);
 
@@ -101,11 +100,12 @@ export const StandardUsage: FunctionComponent = () => {
         <StyledPicker prompt="Avatar Color" selected={avatarColor} onChange={onAvatarColorChange} collection={avatarColors} />
         <StyledPicker prompt="Presence status" selected={presence} onChange={onPresenceChange} collection={allPresences} />
       </View>
-      <JSAvatar badge={{ status: 'available', outOfOffice }} />
-      <JSAvatar name="Richard" avatarColor="#ff0099" initialsColor="yellow" />
-      <JSAvatar icon={{ fontSource: { ...fontBuiltInProps, fontSize: 32 }, color: 'red' }} size={56} />
-      <JSAvatar accessibilityLabel="Fall-back Icon" accessibilityHint="A picture representing a user" size={120} />
-      <JSAvatar
+      <Avatar
+        accessibilityLabel="Fall-back Icon"
+        accessibilityHint="A picture representing a user"
+        size={imageSize === undefinedText ? undefined : imageSize}
+      />
+      <Avatar
         active={active}
         activeAppearance={activeAppearance}
         size={imageSize === undefinedText ? undefined : imageSize}
@@ -116,28 +116,29 @@ export const StandardUsage: FunctionComponent = () => {
         imageUrl={showImage ? satyaPhotoUrl : undefined}
         avatarColor={avatarColor}
       />
-      <JSAvatar
+      <Avatar
         active={active}
         activeAppearance={activeAppearance}
         size={imageSize === undefinedText ? undefined : imageSize}
         shape={isSquare ? 'square' : 'circular'}
         accessibilityLabel="Icon"
         name="* Richard Faynman *"
-        avatarColor={avatarColor}
+        avatarColor="#ff0099"
+        initialsColor="yellow"
       />
       {svgIconsEnabled && (
         <>
-          <JSAvatar
+          <Avatar
             active={active}
             activeAppearance={activeAppearance}
             size={imageSize === undefinedText ? undefined : imageSize}
             shape={isSquare ? 'square' : 'circular'}
             accessibilityLabel="SVG Icon"
-            icon={iconProps}
+            icon={{ fontSource: { ...fontBuiltInProps, fontSize: 32 }, color: 'red' }}
             avatarColor={avatarColor}
             badge={{ status: 'outOfOffice', outOfOffice }}
           />
-          <JSAvatar
+          <Avatar
             accessibilityHint="A picture representing a user"
             active={active}
             activeAppearance={activeAppearance}