+++ packages/components/Avatar/src/AvatarTokens.win32.ts
@@ -1,12 +1,11 @@
 import { Theme } from '@fluentui-react-native/framework';
 import { TokenSettings } from '@fluentui-react-native/use-styling';
-import { JSAvatarTokens } from '.';
+import { AvatarTokens } from '.';
 import { globalTokens } from '@fluentui-react-native/theme-tokens';
 
-export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: Theme) =>
+export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme) =>
   ({
-    horizontalIconAlignment: 'end',
-    verticalIconAlignment: 'end',
+    badgeSize: 'smallest',
     color: t.colors.neutralForeground3,
     backgroundColor: t.colors.neutralBackground6,
     avatarOpacity: 1,
@@ -18,7 +17,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
     iconColor: t.colors.neutralForeground3,
     ringColor: t.colors.transparentStroke,
     borderColor: globalTokens.color.white,
-    borderWidth: t.host.appearance === 'highContrast' ? 1 : 0,
+    borderWidth: t.name === 'HighContrast' ? 1 : 0,
     circular: {
       borderRadius: globalTokens.corner.radius.circle,
     },
@@ -166,7 +165,7 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
     burgundy: getColorProps('burgundy', t),
     hotPink: getColorProps('hotPink', t),
     orchid: getColorProps('orchid', t),
-  } as JSAvatarTokens);
+  } as AvatarTokens);
 
 /**
  * A function which returns object of props depending on color and theme.
@@ -175,9 +174,10 @@ export const defaultJSAvatarTokens: TokenSettings<JSAvatarTokens, Theme> = (t: T
  * @returns object of props - backgroundColor, color and ringColor
  */
 function getColorProps(color: string, theme: Theme) {
-  const themeAppearance = theme.host.appearance;
+  const themeAppearance = theme.name;
   switch (themeAppearance) {
-    case 'light':
+    case 'White':
+    case 'Colorful':
     default:
       return {
         backgroundColor: globalTokens.color[color].tint40,
@@ -185,14 +185,15 @@ function getColorProps(color: string, theme: Theme) {
         iconColor: globalTokens.color[color].shade30,
         ringColor: globalTokens.color[color].primary,
       };
-    case 'dark':
+    case 'DarkGray':
+    case 'Black':
       return {
         backgroundColor: globalTokens.color[color].shade30,
         color: globalTokens.color[color].tint40,
         iconColor: globalTokens.color[color].tint40,
         ringColor: globalTokens.color[color].tint30,
       };
-    case 'highContrast':
+    case 'HighContrast':
       return {
         backgroundColor: theme.colors.neutralBackground6,
         color: theme.colors.neutralForeground3,